backend/requirements.txt
Flask
Flask-Cors
python-dotenv
langchain
langchain_ollama
langchain_google_genai
langchain_community
google-generativeai
protobuf>=6.0.0
requests
pydantic
